+++ api/models/role.go
@@ -4,7 +4,7 @@ import "time"
 
 // Role as Group while on django
 type Role struct {
-	Id          uint64    `json:"id" gorm:"primary_key;AUTO_INCREMENT;not null`
+	Id          int       `json:"id" gorm:"primary_key;AUTO_INCREMENT;not null`
 	Name        string    `json:"name" gorm:"not null;size:20"`
 	CreatedDate time.Time `json:"created_date" gorm:"not null;default:CURRENT_TIMESTAMP;"`
 	UpdateDate  time.Time `json:"update_date" gorm:"not null;default:CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP;"`
